requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow>=9.0.0
pycryptodome==3.18.0
bcrypt==4.0.1
opencv-python-headless==4.8.1.78
WTForms==3.0.1
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
python-magic-bin==0.4.14
numpy>=1.21.0
